feat: add nix development shell
This commit introduces a `shell.nix` file to provide a consistent and
reproducible development environment.

The `shell.nix` file uses `mkShell` to create a shell environment with
Python 3 and pytest available. This ensures that developers have the
necessary dependencies to run tests. This also make sure that the
environment is isolated from the host system, preventing dependency
conflicts.